SatelliteSensor: improve I2C transactions by repeat START
Improve I2C direct sensor polling by combining w/r transactions with
a repeated START since ast2600's i2c new-driver mode can't handle
separated write & read transactions on the i2c-slave-eeprom properly.

Meanwhile, combining with a repeated START can improve performance
and ensure an atomic access (e.g. move offset & read) to the virtual
eeprom.

* Old approach:
  START -> write trans -> STOP -> START -> read trans -> STOP
* New approach:
  START -> write trans -> Repeated START -> read trans -> STOP

Fixes nvbug https://nvbugspro.nvidia.com/bug/5030119

+++ src/SatelliteSensor.hpp
@@ -34,7 +34,8 @@
 #endif
 
 template <typename T>
-int i2cCmd(uint8_t bus, uint8_t addr, size_t offset, T* reading, int length);
+int i2cCmd(uint8_t bus, uint8_t addr, size_t offset, T* reading,
+           uint8_t length);
 
 struct SatelliteSensor : public Sensor
 {
